requirements.txt
absl-py==1.0.0
anyascii==0.3.0
asgiref==3.4.1
astunparse==1.6.3
autopep8==1.4.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bleach==4.1.0
cachetools==4.2.4
certifi==2019.9.11
chardet==3.0.4
click==8.0.3
colorama==0.4.4
dateparser==0.7.2
distlib==0.3.4
Django==4.0.3
django-audit-log==0.7.0
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-filter==21.1
django-model-utils==4.2.0
django-modelcluster==5.3
django-quill-editor==0.1.40
django-role-permissions==2.2.1
django-summernote==0.8.20.0
django-taggit==2.0.0
django-tinymce==3.4.0
django-treebeard==4.5.1
djangorestframework==3.13.1
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
fake-factory==9999.9.9
Faker==2.0.4
filelock==3.4.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
html5lib==1.1
idna==2.8
importlib-metadata==4.10.1
itsdangerous==2.1.1
jet-bridge-base==0.7.7
jet-django==1.0.2
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
keras-nightly==2.9.0.dev2022011708
Keras-Preprocessing==1.1.2
l18n==2021.3
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
mysqlclient==2.1.0
nltk==3.6.7
numpy==1.22.1
oauthlib==3.1.1
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
Pillow==6.2.1
pipenv==2021.11.23
pipreqs==0.4.11
platformdirs==2.4.1
prompt-toolkit==2.0.9
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyparsing==3.0.7
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.3
regex==2021.11.10
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.13.0
soupsieve==2.3.1
SQLAlchemy==1.3.11
sqlparse==0.3.0
tablib==3.2.0
tb-nightly==2.8.0a20220117
telepath==0.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.9.0.dev2022011709
tf-nightly-gpu==2.9.0.dev20220115
tqdm==4.62.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==2.0.0
urllib3==1.25.7
virtualenv==20.11.1
virtualenv-clone==0.5.7
wagtail==2.16.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0.2
Willow==1.4.1
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarg==0.1.9
zipp==3.7.0
